requirements.txt
streamlit>=1.24.0
fpdf2>=2.7.8
markdown2>=2.4.10
beautifulsoup4>=4.12.2
